// src/mocks/types.ts
export type MenuItem = {
  id: string;
  storefrontId: string;
  name: string;
  price: number;
  description: string;
  imageUri?: string;
  categoryId: string;
  availability_status: 'AVAILABLE' | 'SOLD_OUT';
};